+++ flypipe/dependency/__init__.py
@@ -1,2 +1,2 @@
 # ruff: noqa: F401
-from flypipe.dependency.preprocess_mode import PreProcessMode
+from flypipe.dependency.preprocess_mode import PreprocessMode
