+++ flypipe/dependency/preprocess_mode.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
 
-class PreProcessMode(Enum):
+class PreprocessMode(Enum):
     DISABLE = 1  # do not apply the preprocess function (if defined)
-    ACTIVE = 2  # by default run the PreProcess
+    ACTIVE = 2  # by default run the Preprocess
