+++ flypipe/run_context_test.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from flypipe.dependency import PreProcessMode
+from flypipe.dependency.preprocess_mode import PreprocessMode
 from flypipe.node import node
 from flypipe.run_context import RunContext
 
@@ -19,32 +19,32 @@ class TestRunContext:
 
     def test_get_run_preprocess_mode_set_on_nodes_returns_active(self):
         run_context = RunContext(
-            dependencies_preprocess_modes={t2: {t1: PreProcessMode.DISABLE}}
+            dependencies_preprocess_modes={t2: {t1: PreprocessMode.DISABLE}}
         )
 
         assert (
-            run_context.get_run_preprocess_mode().value == PreProcessMode.ACTIVE.value
+            run_context.get_run_preprocess_mode().value == PreprocessMode.ACTIVE.value
         )
 
     def test_get_run_preprocess_mode_not_set_returns_active(self):
         run_context = RunContext()
 
         assert (
-            run_context.get_run_preprocess_mode().value == PreProcessMode.ACTIVE.value
+            run_context.get_run_preprocess_mode().value == PreprocessMode.ACTIVE.value
         )
 
     def test_get_run_preprocess_mode_disabled_returns_disabled(self):
-        run_context = RunContext(dependencies_preprocess_modes=PreProcessMode.DISABLE)
+        run_context = RunContext(dependencies_preprocess_modes=PreprocessMode.DISABLE)
 
         assert (
-            run_context.get_run_preprocess_mode().value == PreProcessMode.DISABLE.value
+            run_context.get_run_preprocess_mode().value == PreprocessMode.DISABLE.value
         )
 
     def test_get_run_preprocess_mode_active_returns_active(self):
-        run_context = RunContext(dependencies_preprocess_modes=PreProcessMode.ACTIVE)
+        run_context = RunContext(dependencies_preprocess_modes=PreprocessMode.ACTIVE)
 
         assert (
-            run_context.get_run_preprocess_mode().value == PreProcessMode.ACTIVE.value
+            run_context.get_run_preprocess_mode().value == PreprocessMode.ACTIVE.value
         )
 
     def test_get_dependency_preprocess_mode_active_if_nothing_is_set(self):
@@ -52,25 +52,25 @@ class TestRunContext:
 
         assert (
             run_context.get_dependency_preprocess_mode(t2, t2.input_nodes[0]).value
-            == PreProcessMode.ACTIVE.value
+            == PreprocessMode.ACTIVE.value
         )
 
     def test_get_dependency_preprocess_mode_disabled_if_disabled(self):
         run_context = RunContext(
-            dependencies_preprocess_modes={t2: {t1: PreProcessMode.DISABLE}}
+            dependencies_preprocess_modes={t2: {t1: PreprocessMode.DISABLE}}
         )
 
         assert (
             run_context.get_dependency_preprocess_mode(t2, t2.input_nodes[0].node).value
-            == PreProcessMode.DISABLE.value
+            == PreprocessMode.DISABLE.value
         )
 
     def test_get_dependency_preprocess_mode_active_if_active(self):
         run_context = RunContext(
-            dependencies_preprocess_modes={t2: {t1: PreProcessMode.ACTIVE}}
+            dependencies_preprocess_modes={t2: {t1: PreprocessMode.ACTIVE}}
         )
 
         assert (
             run_context.get_dependency_preprocess_mode(t2, t2.input_nodes[0].node).value
-            == PreProcessMode.ACTIVE.value
+            == PreprocessMode.ACTIVE.value
         )
